Map house finder
(a very old project, a lot of unnecessary)
A program that implements a search for a house on a map of the area by known geographical coordinates of the corners of the structure and the coordinates of the extreme points of the map image. A graphic highlighting of the desired building on the map is also implemented.